Split MovieWriterPNGWAV into its own file
This makes its organization consistent with MovieWriterMJPEG.
