+++ servers/movie_writer/movie_writer_pngwav.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  movie_writer.h                                                       */
+/*  movie_writer_pngwav.h                                                */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,61 +28,44 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef MOVIE_WRITER_H
-#define MOVIE_WRITER_H
+#ifndef MOVIE_WRITER_PNGWAV_H
+#define MOVIE_WRITER_PNGWAV_H
 
-#include "core/templates/local_vector.h"
-#include "servers/audio/audio_driver_dummy.h"
-#include "servers/audio_server.h"
+#include "servers/movie_writer/movie_writer.h"
 
-class MovieWriter : public Object {
-	GDCLASS(MovieWriter, Object);
-
-	uint64_t fps = 0;
-	uint64_t mix_rate = 0;
-	uint32_t audio_channels = 0;
-
-	LocalVector<int32_t> audio_mix_buffer;
+class MovieWriterPNGWAV : public MovieWriter {
+	GDCLASS(MovieWriterPNGWAV, MovieWriter)
 
 	enum {
-		MAX_WRITERS = 8
+		MAX_TRAILING_ZEROS = 8 // more than 10 days at 60fps, no hard drive can put up with this anyway :)
 	};
-	static MovieWriter *writers[];
-	static uint32_t writer_count;
+
+	uint32_t mix_rate = 48000;
+	AudioServer::SpeakerMode speaker_mode = AudioServer::SPEAKER_MODE_STEREO;
+	String base_path;
+	uint32_t frame_count = 0;
+	uint32_t fps = 0;
+
+	uint32_t audio_block_size = 0;
+
+	Ref<FileAccess> f_wav;
+	uint32_t wav_data_size_pos = 0;
+
+	String zeros_str(uint32_t p_index);
 
 protected:
-	virtual uint32_t get_audio_mix_rate() const;
-	virtual AudioServer::SpeakerMode get_audio_speaker_mode() const;
+	virtual uint32_t get_audio_mix_rate() const override;
+	virtual AudioServer::SpeakerMode get_audio_speaker_mode() const override;
+	virtual void get_supported_extensions(List<String> *r_extensions) const override;
 
-	virtual Error write_begin(const Size2i &p_movie_size, uint32_t p_fps, const String &p_base_path);
-	virtual Error write_frame(const Ref<Image> &p_image, const int32_t *p_audio_data);
-	virtual void write_end();
+	virtual Error write_begin(const Size2i &p_movie_size, uint32_t p_fps, const String &p_base_path) override;
+	virtual Error write_frame(const Ref<Image> &p_image, const int32_t *p_audio_data) override;
+	virtual void write_end() override;
 
-	GDVIRTUAL0RC(uint32_t, _get_audio_mix_rate)
-	GDVIRTUAL0RC(AudioServer::SpeakerMode, _get_audio_speaker_mode)
-
-	GDVIRTUAL1RC(bool, _handles_file, const String &)
-	GDVIRTUAL0RC(Vector<String>, _get_supported_extensions)
-
-	GDVIRTUAL3R(Error, _write_begin, const Size2i &, uint32_t, const String &)
-	GDVIRTUAL2R(Error, _write_frame, const Ref<Image> &, GDNativeConstPtr<int32_t>)
-	GDVIRTUAL0(_write_end)
-
-	static void _bind_methods();
+	virtual bool handles_file(const String &p_path) const override;
 
 public:
-	virtual bool handles_file(const String &p_path) const;
-	virtual void get_supported_extensions(List<String> *r_extensions) const;
-
-	static void add_writer(MovieWriter *p_writer);
-	static MovieWriter *find_writer_for_file(const String &p_file);
-
-	void begin(const Size2i &p_movie_size, uint32_t p_fps, const String &p_base_path);
-	void add_frame(const Ref<Image> &p_image);
-
-	static void set_extensions_hint();
-
-	void end();
+	MovieWriterPNGWAV();
 };
 
-#endif // MOVIE_WRITER_H
+#endif // MOVIE_WRITER_PNGWAV_H
